Support concurrent downloads as per https://github.com/Squirrel/Squirrel.Windows/issues/880 Check if file already downloaded as per https://github.com/cpprobr/Squirrel.Windows/commit/dba74d155f02e1f55d122b3a0d2397161e6ada97

+++ src/Squirrel/UpdateManager.DownloadReleases.cs
@@ -32,6 +32,10 @@ namespace Squirrel
                     // From Internet
                     await releasesToDownload.ForEachAsync(async x => {
                         var targetFile = Path.Combine(packagesDirectory, x.Filename);
+
+                        if (CheckIfAlreadyDownloaded(x, targetFile))
+                            return;
+
                         double component = 0;
                         await downloadRelease(updateUrlOrPath, x, urlDownloader, targetFile, p => {
                             lock (progress) {
@@ -48,6 +52,9 @@ namespace Squirrel
                     await releasesToDownload.ForEachAsync(x => {
                         var targetFile = Path.Combine(packagesDirectory, x.Filename);
 
+                        if (CheckIfAlreadyDownloaded(x, targetFile))
+                            return;
+
                         File.Copy(
                             Path.Combine(updateUrlOrPath, x.Filename),
                             targetFile,
@@ -59,9 +66,23 @@ namespace Squirrel
                 }
             }
 
+            private bool CheckIfAlreadyDownloaded(ReleaseEntry releaseEntry, string targetFile)
+            {
+                if (!File.Exists(targetFile))
+                    return false;
+
+                try {
+                    checksumPackage(releaseEntry);
+                    return true;
+                } catch (Exception) {
+                    // Needs to download again
+                    return false;
+                }
+            }
+
             bool isReleaseExplicitlyHttp(ReleaseEntry x)
             {
-                return x.BaseUrl != null && 
+                return x.BaseUrl != null &&
                     Uri.IsWellFormedUriString(x.BaseUrl, UriKind.Absolute);
             }
 
@@ -105,7 +126,7 @@ namespace Squirrel
                 using (var file = targetPackage.OpenRead()) {
                     var hash = Utility.CalculateStreamSHA1(file);
 
-                    if (!hash.Equals(downloadedRelease.SHA1,StringComparison.OrdinalIgnoreCase)) {
+                    if (!hash.Equals(downloadedRelease.SHA1, StringComparison.OrdinalIgnoreCase)) {
                         this.Log().Error("File SHA1 should be {0}, is {1}", downloadedRelease.SHA1, hash);
                         targetPackage.Delete();
                         throw new Exception("Checksum doesn't match: " + targetPackage.FullName);
